Correct Lua return values, clean up all compiler warnings in BT code
There are still warnings in third-party code (mostly FLU)

Make Flu_Tree_Browser::Node::_id an uintptr_t, use uintptr_t in
Flu_Tree_Browser::{remove, modify} as modify can either return an id
or a pointer

rotater_c: Add virtual destructor

Catch exceptions by reference instead of by value

remove unused variables

Reformat lines with for and while loops to make them more legible

mark all fall throughs in switch statements

Don't assert where we always need to check

xmlWriter_c destructor: mark that it can throw exceptions

+++ src/flu/Flu_Tree_Browser.cpp
@@ -3500,7 +3500,7 @@ Flu_Tree_Browser::Node* Flu_Tree_Browser :: Node :: modify( const char* path, in
 	// if this is the last node, remove it.
 	if( lastNode )
 	  {
-	    int ID = n->id();
+	    uintptr_t ID = n->id();
 	    _children.erase( n );
 	    //if( tree->rdata.cbNode == n )
 	    //tree->rdata.cbNode = NULL;
